+++ src/main/java/com/aston/shop/users/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.aston.shop.users.repository;
 
-import com.aston.shop.users.entity.entity.User;
+import com.aston.shop.users.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
